+++ src/main/java/com/example/community/controller/UserController.java
@@ -3,7 +3,6 @@ package com.example.community.controller;
 import com.example.community.domain.User;
 import com.example.community.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -29,7 +28,7 @@ public class UserController {
   }
 
   @GetMapping("/{id}")
-  public User getUserById(@PathVariable int id) {
+  public User getUserById(@PathVariable Long id) {
     return userService.getUserById(id);
   }
 
@@ -39,7 +38,7 @@ public class UserController {
   }
 
   @DeleteMapping("/{id}")
-  public void deleteUser(@PathVariable int id) {
+  public void deleteUser(@PathVariable Long id) {
     userService.deleteUser(id);
   }
 }
